test: move repository tests to integration test folder
