+++ src/server/repositories/account.rs
@@ -88,57 +88,3 @@ impl Repository {
         Ok(row)
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use anyhow::Context;
-    use anyhow::Result;
-    use sqlx::PgConnection;
-    use sqlx::PgPool;
-
-    async fn create(tx: &mut PgConnection) -> Result<Entity> {
-        let account = Entity::new(
-            testutils::rand::uuid(),
-            testutils::rand::string(10),
-            testutils::rand::email(),
-            testutils::rand::string(10),
-            testutils::rand::string(10),
-            testutils::rand::i64(1, 100000),
-        )
-        .context("failed to validate account")?;
-        Repository::upsert(&account, tx)
-            .await
-            .context("failed to create account")?;
-        Ok(account)
-    }
-
-    #[sqlx::test]
-    #[ignore] // NOTE: Be sure '$ docker compose -f devops/local/docker-compose.yaml up' before running this test
-    async fn test_create_and_select_by_name(pool: PgPool) -> Result<()> {
-        let mut tx = pool
-            .begin()
-            .await
-            .expect("transaction should be started properly");
-        let account = create(&mut tx)
-            .await
-            .expect("new account should be created");
-        let fetched = Repository::select_by_name(account.name(), &mut tx)
-            .await
-            .expect("created account should be found");
-        if let Some(fetched) = fetched {
-            assert_eq!(&fetched.id, account.id().as_uuid());
-            assert_eq!(&fetched.name, account.name().as_str());
-            assert_eq!(&fetched.email, account.email().as_str());
-            assert_eq!(&fetched.password, account.password().as_str());
-            assert_eq!(&fetched.namespace, account.namespace().as_str());
-            assert_eq!(&fetched.ttl, account.ttl().as_i64());
-        } else {
-            panic!("created account should be matched");
-        }
-        tx.rollback()
-            .await
-            .expect("rollback should be done properly");
-        Ok(())
-    }
-}
