+++ src/server/repositories/schema.rs
@@ -85,90 +85,3 @@ impl Repository {
         Ok(row)
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use anyhow::Context;
-    use anyhow::Result;
-    use sqlx::PgConnection;
-    use sqlx::PgPool;
-
-    use super::*;
-    use crate::server::entities::account::Entity as Account;
-    use crate::server::entities::account::Id as AccountId;
-    use crate::server::entities::share::Entity as Share;
-    use crate::server::repositories::account::Repository as AccountRepository;
-    use crate::server::repositories::share::Repository as ShareRepository;
-
-    async fn create_account(tx: &mut PgConnection) -> Result<Account> {
-        let account = Account::new(
-            testutils::rand::uuid(),
-            testutils::rand::string(10),
-            testutils::rand::email(),
-            testutils::rand::string(10),
-            testutils::rand::string(10),
-            testutils::rand::i64(1, 100000),
-        )
-        .context("failed to validate account")?;
-        AccountRepository::upsert(&account, tx)
-            .await
-            .context("failed to create account")?;
-        Ok(account)
-    }
-
-    async fn create_share(account_id: &AccountId, tx: &mut PgConnection) -> Result<Share> {
-        let share = Share::new(
-            testutils::rand::uuid(),
-            testutils::rand::string(10),
-            account_id.to_uuid().to_string(),
-        )
-        .context("failed to validate share")?;
-        ShareRepository::upsert(&share, tx)
-            .await
-            .context("failed to create share")?;
-        Ok(share)
-    }
-
-    async fn create_schema(
-        account_id: &AccountId,
-        share_id: &ShareId,
-        tx: &mut PgConnection,
-    ) -> Result<Entity> {
-        let schema = Entity::new(
-            testutils::rand::uuid(),
-            testutils::rand::string(10),
-            share_id.to_uuid().to_string(),
-            account_id.to_uuid().to_string(),
-        )
-        .context("failed to validate schema")?;
-        Repository::upsert(&schema, tx)
-            .await
-            .context("failed to create schema")?;
-        Ok(schema)
-    }
-
-    #[sqlx::test]
-    #[ignore] // NOTE: Be sure '$ docker compose -f devops/local/docker-compose.yaml up' before running this test
-    async fn test_create(pool: PgPool) -> Result<()> {
-        let mut tx = pool
-            .begin()
-            .await
-            .expect("transaction should be started properly");
-        let account = create_account(&mut tx)
-            .await
-            .expect("new account should be created");
-        let share = create_share(account.id(), &mut tx)
-            .await
-            .expect("new share should be created");
-        let records = testutils::rand::i64(0, 20);
-        for _ in 0..records {
-            create_schema(account.id(), share.id(), &mut tx)
-                .await
-                .expect("new schema should be created");
-        }
-        tx.rollback()
-            .await
-            .expect("rollback should be done properly");
-        Ok(())
-    }
-}
