+++ src/server/repositories/share.rs
@@ -73,73 +73,3 @@ impl Repository {
         Ok(row)
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::server::entities::account::Entity as Account;
-    use crate::server::entities::account::Id as AccountId;
-    use crate::server::repositories::account::Repository as AccountRepository;
-    use anyhow::Context;
-    use anyhow::Result;
-    use sqlx::PgConnection;
-    use sqlx::PgPool;
-
-    async fn create_account(tx: &mut PgConnection) -> Result<Account> {
-        let account = Account::new(
-            testutils::rand::uuid(),
-            testutils::rand::string(10),
-            testutils::rand::email(),
-            testutils::rand::string(10),
-            testutils::rand::string(10),
-            testutils::rand::i64(1, 100000),
-        )
-        .context("failed to validate account")?;
-        AccountRepository::upsert(&account, tx)
-            .await
-            .context("failed to create account")?;
-        Ok(account)
-    }
-
-    async fn create_share(account_id: &AccountId, tx: &mut PgConnection) -> Result<Entity> {
-        let share = Entity::new(
-            testutils::rand::uuid(),
-            testutils::rand::string(10),
-            account_id.to_uuid().to_string(),
-        )
-        .context("failed to validate share")?;
-        Repository::upsert(&share, tx)
-            .await
-            .context("failed to create share")?;
-        Ok(share)
-    }
-
-    #[sqlx::test]
-    #[ignore] // NOTE: Be sure '$ docker compose -f devops/local/docker-compose.yaml up' before running this test
-    async fn test_create_and_select_by_name(pool: PgPool) -> Result<()> {
-        let mut tx = pool
-            .begin()
-            .await
-            .expect("transaction should be started properly");
-        let account = create_account(&mut tx)
-            .await
-            .expect("new account should be created");
-        let share = create_share(account.id(), &mut tx)
-            .await
-            .expect("new share should be created");
-        let fetched = Repository::select_by_name(share.name(), &mut tx)
-            .await
-            .expect("created share should be found");
-        if let Some(fetched) = fetched {
-            assert_eq!(&fetched.id, share.id().as_uuid());
-            assert_eq!(&fetched.name, share.name().as_str());
-            assert_eq!(&fetched.created_by, share.created_by().as_uuid());
-        } else {
-            panic!("created share should be matched");
-        }
-        tx.rollback()
-            .await
-            .expect("rollback should be done properly");
-        Ok(())
-    }
-}
