+++ src/server/repositories/token.rs
@@ -56,69 +56,3 @@ impl Repository {
         ))
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::server::entities::account::Entity as Account;
-    use crate::server::entities::account::Id as AccountId;
-    use crate::server::repositories::account::Repository as AccountRepository;
-    use anyhow::Context;
-    use anyhow::Result;
-    use sqlx::PgConnection;
-    use sqlx::PgPool;
-    use std::str::FromStr;
-
-    async fn create_account(tx: &mut PgConnection) -> Result<Account> {
-        let account = Account::new(
-            testutils::rand::uuid(),
-            testutils::rand::string(10),
-            testutils::rand::email(),
-            testutils::rand::string(10),
-            testutils::rand::string(10),
-            testutils::rand::i64(1, 100000),
-        )
-        .context("failed to validate account")?;
-        AccountRepository::upsert(&account, tx)
-            .await
-            .context("failed to create account")?;
-        Ok(account)
-    }
-
-    async fn create_token(account_id: &AccountId, tx: &mut PgConnection) -> Result<Entity> {
-        let roles = vec!["Admin", "Guest"];
-        let role = testutils::rand::choose(&roles);
-        let role = Role::from_str(role).context("failed to choose role")?;
-        let token = Entity::new(
-            testutils::rand::uuid(),
-            testutils::rand::email(),
-            role,
-            testutils::rand::string(10),
-            account_id.to_uuid().to_string(),
-        )
-        .context("failed to validate token")?;
-        Repository::upsert(&token, tx)
-            .await
-            .context("failed to create token")?;
-        Ok(token)
-    }
-
-    #[sqlx::test]
-    #[ignore] // NOTE: Be sure '$ docker compose -f devops/local/docker-compose.yaml up' before running this test
-    async fn test_create(pool: PgPool) -> Result<()> {
-        let mut tx = pool
-            .begin()
-            .await
-            .expect("transaction should be started properly");
-        let account = create_account(&mut tx)
-            .await
-            .expect("new account should be created");
-        create_token(account.id(), &mut tx)
-            .await
-            .expect("new token should be created");
-        tx.rollback()
-            .await
-            .expect("rollback should be done properly");
-        Ok(())
-    }
-}
